feat(core)!: align agent, streaming and modality lifecycle contracts
Align agent preparation, per-step overrides, callback ordering, tool parsing, output aggregation, independent stream views and modality lifecycle events. Preserve cancellation and telemetry content filtering; add corresponding regression tests and API snapshots.

BREAKING CHANGE: Embeddings use f64; Instructions is an enum; prepared-step model access exposes the model instance; telemetry callbacks return BoxFuture; result accessors and event structures have changed. See the unreleased migration guide and ADR 0026.

+++ crates/ferrin-core/tests/suite/modalities/common.rs
@@ -120,10 +120,10 @@ impl EmbeddingModel for EmbedMock {
             }
             let count = options.values.len();
             #[allow(clippy::cast_precision_loss, reason = "test values are tiny")]
-            let mut embeddings: Vec<Vec<f32>> = options
+            let mut embeddings: Vec<Vec<f64>> = options
                 .values
                 .iter()
-                .map(|value| vec![value.len() as f32, 1.0])
+                .map(|value| vec![value.len() as f64, 1.0])
                 .collect();
             embeddings.truncate(count.saturating_sub(drop));
             Ok(EmbedResult {
